old compiler and os fix (whitespace-only changes removed)
This commit is based on commit 3056c18 "old compiler and os fix" but
with whitespace-only changes removed. Only substantive changes are
preserved:

- Type changes: long long -> int64_t, __int64
- Literal suffix changes: ULL -> UI64, LL -> I64
- Platform-specific compatibility fixes (Windows, VMS, etc.)
- Code blocks wrapped in #if 0 (test/json_test.c)
- New includes and function pointer handling

Whitespace-only changes removed:
- Cast spacing: (type) func -> (type)func
- Function call spacing: func (...) -> func(...)
- Indentation-only changes in for loops and comments
- Formatting changes in macro definitions

All comments preserved as-is.

+++ test/radix/terp.c
@@ -421,7 +421,7 @@ static int SRDR_print_one(SRDR *srdr, BIO *bio, size_t i, int *was_end)
         GET_OPERAND(srdr, v);
         PRINT_OPC(PUSH_U64);
         BIO_printf(bio, "%#20llx (%llu)",
-            (unsigned long long)v, (unsigned long long)v);
+            (unsigned __int64)v, (unsigned __int64)v);
     } break;
     case OPK_PUSH_SIZE: {
         size_t v;
@@ -429,7 +429,7 @@ static int SRDR_print_one(SRDR *srdr, BIO *bio, size_t i, int *was_end)
         GET_OPERAND(srdr, v);
         PRINT_OPC(PUSH_SIZE);
         BIO_printf(bio, "%#20llx (%llu)",
-            (unsigned long long)v, (unsigned long long)v);
+            (unsigned __int64)v, (unsigned __int64)v);
     } break;
     case OPK_FUNC: {
         helper_func_t v;
@@ -452,7 +452,7 @@ static int SRDR_print_one(SRDR *srdr, BIO *bio, size_t i, int *was_end)
     } break;
     default:
         TEST_error("unsupported opcode while printing: %llu",
-            (unsigned long long)opc);
+            (unsigned __int64)opc);
         goto err;
     }
 
@@ -518,7 +518,8 @@ static void SCRIPT_INFO_print(SCRIPT_INFO *script_info, BIO *bio, int error,
 typedef struct terp_config_st {
     BIO *debug_bio;
 
-    OSSL_TIME (*now_cb)(void *arg);
+    OSSL_TIME (*now_cb)
+    (void *arg);
     void *now_cb_arg;
 
     int (*per_op_cb)(TERP *terp, void *arg);
@@ -774,7 +775,7 @@ static int TERP_execute(TERP *terp)
             }
         } break;
         default:
-            TEST_error("unknown opcode: %llu", (unsigned long long)opc);
+            TEST_error("unknown opcode: %llu", (unsigned __int64)opc);
             goto err;
         }
     }
@@ -850,7 +851,7 @@ err:
     GEN_SCRIPT_cleanup(&gen_script);
     if (have_terp) {
         BIO_printf(debug_bio, "Stats:\n  Ops executed: %16llu\n\n",
-            (unsigned long long)terp.ops_executed);
+            (unsigned __int64)terp.ops_executed);
     }
     SCRIPT_INFO_print(script_info, debug_bio, /*error=*/!ok,
         ok ? "completed" : "failed, exiting");
